OK, maybe I can work with this.
All sprites HAVE to be scaled the same. The players are 128x128 and
drawn with a bit of a S c r o n t c h but they mostly line up.

One of the issues before was that the dirt block is 164x164, which is
not the same. I think if I force a "block" to be 128x128, and keep the
scaling factors and everything uniform, then we can
make this work. Proceeding with caution.