Disable trap (for now)
I really wanna finish this game, and it's been SO long since I've worked
on it. I figure this is enough of a tech demo to publish, and it's been
long enough since I've cut a release of Wraith Hunt for people to play,
so while I iron out the next round of issues, here goes nothing!